requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dropbox==10.5.0
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
oauth2client==4.1.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.3
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
